En la Imprenta
Added images to add visual appeal. I tried to find interesting ones showcasing the team winning, the robot, or the drive team.
The one for 2015 may change. I chose it because Woodie's in it, but due to its height, it leaves a slightly unsightly gap.
Other formatting properties subject to change.