In the Press long-needed change
Finally got around to doing this with the no-hw weekend.

Took inspiration from the Machining2 page. This should work better, as there is very little content.
Space has been filled so it looks much nicer.
Black bar (empty <di>) still exists. Filled it with generic text.
Format may change (hopefully to make it look even better); if this is unsatisfactory, it can be reverted to the original easily to to the original's simplicity.